feat: Style the widget section
Followed the CSS-tricks suggestions on
https://css-tricks.com/snippets/css/form-validation-styling-on-input-focus/
to only mark the input text fields as invalid when text is entered by
the user. However this approach is still not optimal. It relies on the
:place-holder-shown pseudo class. If the user focuses any field which is
required and moves on without entering any text the placeholder is shown
again and the text field is not marked as invalid. For this to show the
user must first submit the form.